// frontend/src/api/generated/models/allocation-order-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AllocationOrderSummary
 */
export interface AllocationOrderSummary {
    /**
     * 出荷伝票ID
     * @type {number}
     * @memberof AllocationOrderSummary
     */
    'id'?: number;
    /**
     * 伝票番号
     * @type {string}
     * @memberof AllocationOrderSummary
     */
    'slipNumber'?: string;
    /**
     * 出荷予定日
     * @type {string}
     * @memberof AllocationOrderSummary
     */
    'plannedDate'?: string;
    /**
     * 出荷先取引先名
     * @type {string}
     * @memberof AllocationOrderSummary
     */
    'partnerName'?: string;
    /**
     * 明細件数
     * @type {number}
     * @memberof AllocationOrderSummary
     */
    'lineCount'?: number;
    /**
     * 受注ステータス
     * @type {string}
     * @memberof AllocationOrderSummary
     */
    'status'?: AllocationOrderSummaryStatusEnum;
}

export const AllocationOrderSummaryStatusEnum = {
    Ordered: 'ORDERED',
    PartialAllocated: 'PARTIAL_ALLOCATED'
} as const;

export type AllocationOrderSummaryStatusEnum = typeof AllocationOrderSummaryStatusEnum[keyof typeof AllocationOrderSummaryStatusEnum];